Reauthenticate when websocket returns forbidden
When a websocket is closed with a "Forbidden" reason, the client redirects to a new /login endpoint.  This bypasses the service worker and hits the authentication logic on the server.  Following successful authentication, the user is redirected back to the relevant path.  (If the user is still authenticated in lucos_auth, this happens entirely seemlessly, assuming reliable connectivity).

Whilst it could be possible to port more of the authenication logic to happen client side, there'll always need to be a server call at some point, so its perferable to avoid duplication of logic.  The user is likely to have network connectivity as the "Forbidden" error in the websocket, which triggers this journey, can only be received if there's network connectivity.